refactor: extract capability actions from AgentWorkspace
