+++ src/components/agent/capability-actions.ts
@@ -1,0 +1,108 @@
+import {
+  toggleHermesBridgeSkill,
+  toggleHermesBridgeToolset,
+  updateHermesBridgeMessagingPlatform,
+  type HermesMessagingPlatformInfo,
+  type HermesSkillInfo,
+  type HermesToolsetInfo,
+} from "../../lib/tauri";
+import { messageFromError } from "../../lib/errors";
+import type { createCapabilityActionsDependencies } from "./capability-actions-types";
+
+export function createCapabilityActions(dependencies: createCapabilityActionsDependencies) {
+  const {
+    loadMessagingPlatforms,
+    messagingEnvEdits,
+    setCapabilitySaving,
+    setError,
+    setMessagingEnvEdits,
+    setMessagingPlatforms,
+    setSkills,
+    setToolsets,
+  } = dependencies;
+
+  async function setSkillEnabled(skill: HermesSkillInfo, enabled: boolean) {
+    setCapabilitySaving(`skill:${skill.name}`);
+    try {
+      await toggleHermesBridgeSkill({ name: skill.name, enabled });
+      setSkills(
+        (current) =>
+          current?.map((item) => (item.name === skill.name ? { ...item, enabled } : item)) ??
+          current,
+      );
+    } catch (err) {
+      setError(messageFromError(err));
+    } finally {
+      setCapabilitySaving(null);
+    }
+  }
+
+  async function setToolsetEnabled(toolset: HermesToolsetInfo, enabled: boolean) {
+    setCapabilitySaving(`toolset:${toolset.name}`);
+    try {
+      await toggleHermesBridgeToolset({ name: toolset.name, enabled });
+      setToolsets(
+        (current) =>
+          current?.map((item) => (item.name === toolset.name ? { ...item, enabled } : item)) ??
+          current,
+      );
+    } catch (err) {
+      setError(messageFromError(err));
+    } finally {
+      setCapabilitySaving(null);
+    }
+  }
+
+  async function setMessagingPlatformEnabled(
+    platform: HermesMessagingPlatformInfo,
+    enabled: boolean,
+  ) {
+    setCapabilitySaving(`messaging:${platform.id}`);
+    try {
+      await updateHermesBridgeMessagingPlatform({
+        platformId: platform.id,
+        enabled,
+      });
+      setMessagingPlatforms(
+        (current) =>
+          current?.map((item) => (item.id === platform.id ? { ...item, enabled } : item)) ??
+          current,
+      );
+    } catch (err) {
+      setError(messageFromError(err));
+    } finally {
+      setCapabilitySaving(null);
+    }
+  }
+
+  async function saveMessagingPlatformEnv(platform: HermesMessagingPlatformInfo) {
+    const env = Object.fromEntries(
+      Object.entries(messagingEnvEdits)
+        .map(([key, value]) => [key, value.trim()])
+        .filter(([, value]) => value.length > 0),
+    );
+    if (!Object.keys(env).length) {
+      return;
+    }
+    setCapabilitySaving(`env:${platform.id}`);
+    try {
+      await updateHermesBridgeMessagingPlatform({
+        platformId: platform.id,
+        env,
+      });
+      setMessagingEnvEdits({});
+      await loadMessagingPlatforms();
+    } catch (err) {
+      setError(messageFromError(err));
+    } finally {
+      setCapabilitySaving(null);
+    }
+  }
+
+  return {
+    setSkillEnabled,
+    setToolsetEnabled,
+    setMessagingPlatformEnabled,
+    saveMessagingPlatformEnv,
+  };
+}
